Workshop, public parties, and multiple player builds

+++ src/main/java/tanks/obstacle/ObstacleTeleporter.java
@@ -198,7 +198,7 @@ public class ObstacleTeleporter extends Obstacle
 		if (s.isEmpty())
 			s = "0";
 
-		this.groupID = Integer.parseInt(s);
+		this.groupID = (int) Double.parseDouble(s);
 		this.refreshMetadata();
 	}
